feat: Consolidate lookup types and source types into a single constants file
