+++ src/common/constants/lookup.constants.ts
@@ -747,3 +747,13 @@ export const lookupConfig: Record<string, any> = {
     ],
   },
 };
+
+export const LOOKUP_TYPES = {
+  AUTHORITIES: 'authorities',
+  HUBS: 'hubs',
+} as const;
+
+export const SOURCE_TYPES = {
+  LOCAL: 'local',
+  LIBRARY_OF_CONGRESS: 'libraryOfCongress',
+} as const;
