+++ src/features/complexLookup/hooks/useAuthoritiesAssignment.ts
@@ -1,11 +1,10 @@
 import { useCallback, useState } from 'react';
 
+import { LOOKUP_TYPES } from '@/common/constants/lookup.constants';
 import { StatusType } from '@/common/constants/status.constants';
 import { AdvancedFieldType } from '@/common/constants/uiControls.constants';
 import { useServicesContext } from '@/common/hooks/useServicesContext';
 import { UserNotificationFactory } from '@/common/services/userNotification';
-
-import { LOOKUP_TYPES } from '@/features/complexLookup/constants/complexLookup.constants';
 
 import { useInputsState, useMarcPreviewState, useProfileState, useStatusState } from '@/store';
