+++ src/features/complexLookup/types/complexLookup.types.ts
@@ -1,3 +1,3 @@
-import { SOURCE_TYPES } from '../constants/complexLookup.constants';
+import { SOURCE_TYPES } from '@/common/constants/lookup.constants';
 
 export type SourceType = (typeof SOURCE_TYPES)[keyof typeof SOURCE_TYPES];
