+++ src/features/complexLookup/utils/complexLookup.helper.ts
@@ -1,10 +1,10 @@
+import { SOURCE_TYPES } from '@/common/constants/lookup.constants';
 import { AdvancedFieldType } from '@/common/constants/uiControls.constants';
 
 import {
   AuthorityValidationTarget,
   COMPLEX_LOOKUPS_LINKED_FIELDS_MAPPING,
   EMPTY_LINKED_DROPDOWN_OPTION_SUFFIX,
-  SOURCE_TYPES,
   VALUE_DIVIDER,
 } from '@/features/complexLookup/constants/complexLookup.constants';
 import { SourceType } from '@/features/complexLookup/types';
